La responsabilidad de guardar unqfy en la db, pasó al repositorio y lo maneja junto con el filesystem
