+++ unqfy/unqfy.js
@@ -1,13 +1,9 @@
-const picklify = require('picklify'); // para cargar/guarfar unqfy
-const fs = require('fs'); // para cargar/guarfar unqfy
 const flatMap = require('array.prototype.flatmap');
 const IdGenerator = require('../model/idGenerator');
 
 const Artist = require('../model/artist');
 const Album = require('../model/album');
 const Track = require('../model/track');
-const Playlist = require('../model/playlist');
-const PlaylistService = require('../unqfy/playlistService');
 
 class UNQfy {
 
@@ -144,22 +140,6 @@ class UNQfy {
     getTracksMatchingArtist(artistName) {
         let artist = this.getArtistByName(artistName);
         return artist.getTracks();
-    }
-
-    save(filename) {
-        const listenersBkp = this.listeners;
-        this.listeners = [];
-
-        const serializedData = picklify.picklify(this);
-
-        this.listeners = listenersBkp;
-        fs.writeFileSync(filename, JSON.stringify(serializedData, null, 2));
-    }
-
-    static load(filename) {
-        const serializedData = fs.readFileSync(filename, { encoding: 'utf-8' });
-        const classes = [UNQfy, Artist, Album, Track, Playlist, PlaylistService];
-        return picklify.unpicklify(JSON.parse(serializedData), classes);
     }
 
     addAnArtist(artist) {
